Move tileset stuff into separate file
